refactor: move jitter backoff generator to utils
